refactor(bulk_metadata_compliance): move method bulk_load_compliance_frameworks inside Compliance class
